requirements.txt
pypdf
langchain
langchain-community
sentence-transformers 
transformers
tqdm
faiss-cpu
pinecone
python-dotenv
ipywidgets
protobuf
bitsandbytes==0.46.1
timm
sentencepiece
tokenizers==0.21.4
accelerate==1.9.0
peft==0.16.0
trl
mistralai
streamlit